Refactor deposit and meal handling, update data models and API responses

+++ app/src/main/java/com/logicline/mydining/models/MessUser.kt
@@ -3,10 +3,10 @@ import com.google.gson.annotations.SerializedName
 import java.time.LocalDateTime
 
 data class MessUser(
-    @SerializedName("id") val id: Long,
-    @SerializedName("mess_id") val messId: Long,
-    @SerializedName("user_id") val userId: Long,
-    @SerializedName("mess_role_id") val messRoleId: Long,
+    @SerializedName("id") val id: Int,
+    @SerializedName("mess_id") val messId: Int,
+    @SerializedName("user_id") val userId: Int,
+    @SerializedName("mess_role_id") val messRoleId: Int,
     @SerializedName("joined_at") val joinedAt: LocalDateTime,
     @SerializedName("left_at") val leftAt: LocalDateTime?,
     @SerializedName("status") val status: String,
